ci-jecbi: fix formatAge NaN guard, extract shared util, fix vitest act() errors
