+++ cmd/ct/assets/web/assets/index-bUUc_DK2.js
@@ -65,4 +65,4 @@ Error generating stack: `+i.message+`
  * LICENSE.md file in the root directory of this source tree.
  *
  * @license MIT
- */function pr(){return pr=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},pr.apply(this,arguments)}function Zd(e,t){if(e==null)return{};var n={},r=Object.keys(e),l,i;for(i=0;i<r.length;i++)l=r[i],!(t.indexOf(l)>=0)&&(n[l]=e[l]);return n}function cg(e){return!!(e.metaKey||e.altKey||e.ctrlKey||e.shiftKey)}function fg(e,t){return e.button===0&&(!t||t==="_self")&&!cg(e)}const dg=["onClick","relative","reloadDocument","replace","state","target","to","preventScrollReset","viewTransition"],pg=["aria-current","caseSensitive","className","end","style","to","viewTransition","children"],hg="6";try{window.__reactRouterVersion=hg}catch{}function mg(e,t){return kv({basename:void 0,future:pr({},void 0,{v7_prependBasename:!0}),history:Km({window:void 0}),hydrationData:vg(),routes:e,mapRouteProperties:sg,dataStrategy:void 0,patchRoutesOnNavigation:void 0,window:void 0}).initialize()}function vg(){var e;let t=(e=window)==null?void 0:e.__staticRouterHydrationData;return t&&t.errors&&(t=pr({},t,{errors:gg(t.errors)})),t}function gg(e){if(!e)return null;let t=Object.entries(e),n={};for(let[r,l]of t)if(l&&l.__type==="RouteErrorResponse")n[r]=new Li(l.status,l.statusText,l.data,l.internal===!0);else if(l&&l.__type==="Error"){if(l.__subType){let i=window[l.__subType];if(typeof i=="function")try{let o=new i(l.message);o.stack="",n[r]=o}catch{}}if(n[r]==null){let i=new Error(l.message);i.stack="",n[r]=i}}else n[r]=l;return n}const qd=P.createContext({isTransitioning:!1}),yg=P.createContext(new Map),wg="startTransition",vc=Mp[wg],xg="flushSync",gc=Qm[xg];function Sg(e){vc?vc(e):e()}function Dr(e){gc?gc(e):e()}class Eg{constructor(){this.status="pending",this.promise=new Promise((t,n)=>{this.resolve=r=>{this.status==="pending"&&(this.status="resolved",t(r))},this.reject=r=>{this.status==="pending"&&(this.status="rejected",n(r))}})}}function kg(e){let{fallbackElement:t,router:n,future:r}=e,[l,i]=P.useState(n.state),[o,a]=P.useState(),[u,s]=P.useState({isTransitioning:!1}),[d,f]=P.useState(),[p,x]=P.useState(),[E,k]=P.useState(),T=P.useRef(new Map),{v7_startTransition:h}=r||{},c=P.useCallback(_=>{h?Sg(_):_()},[h]),v=P.useCallback((_,$)=>{let{deletedFetchers:D,flushSync:q,viewTransitionOpts:re}=$;_.fetchers.forEach((Pe,ut)=>{Pe.data!==void 0&&T.current.set(ut,Pe.data)}),D.forEach(Pe=>T.current.delete(Pe));let xe=n.window==null||n.window.document==null||typeof n.window.document.startViewTransition!="function";if(!re||xe){q?Dr(()=>i(_)):c(()=>i(_));return}if(q){Dr(()=>{p&&(d&&d.resolve(),p.skipTransition()),s({isTransitioning:!0,flushSync:!0,currentLocation:re.currentLocation,nextLocation:re.nextLocation})});let Pe=n.window.document.startViewTransition(()=>{Dr(()=>i(_))});Pe.finished.finally(()=>{Dr(()=>{f(void 0),x(void 0),a(void 0),s({isTransitioning:!1})})}),Dr(()=>x(Pe));return}p?(d&&d.resolve(),p.skipTransition(),k({state:_,currentLocation:re.currentLocation,nextLocation:re.nextLocation})):(a(_),s({isTransitioning:!0,flushSync:!1,currentLocation:re.currentLocation,nextLocation:re.nextLocation}))},[n.window,p,d,T,c]);P.useLayoutEffect(()=>n.subscribe(v),[n,v]),P.useEffect(()=>{u.isTransitioning&&!u.flushSync&&f(new Eg)},[u]),P.useEffect(()=>{if(d&&o&&n.window){let _=o,$=d.promise,D=n.window.document.startViewTransition(async()=>{c(()=>i(_)),await $});D.finished.finally(()=>{f(void 0),x(void 0),a(void 0),s({isTransitioning:!1})}),x(D)}},[c,o,d,n.window]),P.useEffect(()=>{d&&o&&l.location.key===o.location.key&&d.resolve()},[d,p,l.location,o]),P.useEffect(()=>{!u.isTransitioning&&E&&(a(E.state),s({isTransitioning:!0,flushSync:!1,currentLocation:E.currentLocation,nextLocation:E.nextLocation}),k(void 0))},[u.isTransitioning,E]),P.useEffect(()=>{},[]);let C=P.useMemo(()=>({createHref:n.createHref,encodeLocation:n.encodeLocation,go:_=>n.navigate(_),push:(_,$,D)=>n.navigate(_,{state:$,preventScrollReset:D==null?void 0:D.preventScrollReset}),replace:(_,$,D)=>n.navigate(_,{replace:!0,state:$,preventScrollReset:D==null?void 0:D.preventScrollReset})}),[n]),j=n.basename||"/",z=P.useMemo(()=>({router:n,navigator:C,static:!1,basename:j}),[n,C,j]),g=P.useMemo(()=>({v7_relativeSplatPath:n.future.v7_relativeSplatPath}),[n.future.v7_relativeSplatPath]);return P.useEffect(()=>og(r,n.future),[r,n.future]),P.createElement(P.Fragment,null,P.createElement(yl.Provider,{value:z},P.createElement(ju.Provider,{value:l},P.createElement(yg.Provider,{value:T.current},P.createElement(qd.Provider,{value:u},P.createElement(ug,{basename:j,location:l.location,navigationType:l.historyAction,navigator:C,future:g},l.initialized||n.future.v7_partialHydration?P.createElement(Cg,{routes:n.routes,future:n.future,state:l}):t))))),null)}const Cg=P.memo(Pg);function Pg(e){let{routes:t,future:n,state:r}=e;return Xv(t,void 0,r,n)}const _g=typeof window<"u"&&typeof window.document<"u"&&typeof window.document.createElement<"u",Ng=/^(?:[a-z][a-z0-9+.-]*:|\/\/)/i,Rg=P.forwardRef(function(t,n){let{onClick:r,relative:l,reloadDocument:i,replace:o,state:a,target:u,to:s,preventScrollReset:d,viewTransition:f}=t,p=Zd(t,dg),{basename:x}=P.useContext(dn),E,k=!1;if(typeof s=="string"&&Ng.test(s)&&(E=s,_g))try{let v=new URL(window.location.href),C=s.startsWith("//")?new URL(v.protocol+s):new URL(s),j=Ft(C.pathname,x);C.origin===v.origin&&j!=null?s=j+C.search+C.hash:k=!0}catch{}let T=Hv(s,{relative:l}),h=Tg(s,{replace:o,state:a,target:u,preventScrollReset:d,relative:l,viewTransition:f});function c(v){r&&r(v),v.defaultPrevented||h(v)}return P.createElement("a",pr({},p,{href:E||T,onClick:k||i?r:c,ref:n,target:u}))}),jg=P.forwardRef(function(t,n){let{"aria-current":r="page",caseSensitive:l=!1,className:i="",end:o=!1,style:a,to:u,viewTransition:s,children:d}=t,f=Zd(t,pg),p=Xi(u,{relative:f.relative}),x=gr(),E=P.useContext(ju),{navigator:k,basename:T}=P.useContext(dn),h=E!=null&&Dg(p)&&s===!0,c=k.encodeLocation?k.encodeLocation(p).pathname:p.pathname,v=x.pathname,C=E&&E.navigation&&E.navigation.location?E.navigation.location.pathname:null;l||(v=v.toLowerCase(),C=C?C.toLowerCase():null,c=c.toLowerCase()),C&&T&&(C=Ft(C,T)||C);const j=c!=="/"&&c.endsWith("/")?c.length-1:c.length;let z=v===c||!o&&v.startsWith(c)&&v.charAt(j)==="/",g=C!=null&&(C===c||!o&&C.startsWith(c)&&C.charAt(c.length)==="/"),_={isActive:z,isPending:g,isTransitioning:h},$=z?r:void 0,D;typeof i=="function"?D=i(_):D=[i,z?"active":null,g?"pending":null,h?"transitioning":null].filter(Boolean).join(" ");let q=typeof a=="function"?a(_):a;return P.createElement(Rg,pr({},f,{"aria-current":$,className:D,ref:n,style:q,to:u,viewTransition:s}),typeof d=="function"?d(_):d)});var Ra;(function(e){e.UseScrollRestoration="useScrollRestoration",e.UseSubmit="useSubmit",e.UseSubmitFetcher="useSubmitFetcher",e.UseFetcher="useFetcher",e.useViewTransitionState="useViewTransitionState"})(Ra||(Ra={}));var yc;(function(e){e.UseFetcher="useFetcher",e.UseFetchers="useFetchers",e.UseScrollRestoration="useScrollRestoration"})(yc||(yc={}));function Lg(e){let t=P.useContext(yl);return t||K(!1),t}function Tg(e,t){let{target:n,replace:r,state:l,preventScrollReset:i,relative:o,viewTransition:a}=t===void 0?{}:t,u=Wv(),s=gr(),d=Xi(e,{relative:o});return P.useCallback(f=>{if(fg(f,n)){f.preventDefault();let p=r!==void 0?r:Mn(s)===Mn(d);u(e,{replace:p,state:l,preventScrollReset:i,relative:o,viewTransition:a})}},[s,u,d,r,l,n,e,i,o,a])}function Dg(e,t){t===void 0&&(t={});let n=P.useContext(qd);n==null&&K(!1);let{basename:r}=Lg(Ra.useViewTransitionState),l=Xi(e,{relative:t.relative});if(!n.isTransitioning)return!1;let i=Ft(n.currentLocation.pathname,r)||n.currentLocation.pathname,o=Ft(n.nextLocation.pathname,r)||n.nextLocation.pathname;return ji(l.pathname,o)!=null||ji(l.pathname,i)!=null}const Mg=[{to:"/app/",label:"Dashboard",icon:Og},{to:"/app/droplets",label:"Droplets",icon:Fg},{to:"/app/castellarius",label:"Castellarius",icon:Ig},{to:"/app/doctor",label:"Doctor",icon:Ug},{to:"/app/logs",label:"Logs",icon:Ag},{to:"/app/repos",label:"Repos/Skills",icon:Bg}];function zg({open:e,onToggle:t}){const n=gr();return y.jsxs(y.Fragment,{children:[e&&y.jsx("div",{className:"fixed inset-0 bg-black/50 z-20 md:hidden",onClick:t}),y.jsxs("aside",{className:`fixed md:relative z-30 h-full bg-cistern-surface border-r border-cistern-border flex flex-col transition-all duration-200 ${e?"w-56":"w-0 md:w-14"} overflow-hidden`,children:[y.jsxs("div",{className:"flex items-center gap-3 px-4 py-4 border-b border-cistern-border min-h-[60px]",children:[y.jsx("button",{onClick:t,className:"text-cistern-muted hover:text-cistern-fg transition-colors","aria-label":"Toggle sidebar",children:y.jsx("svg",{width:"20",height:"20",viewBox:"0 0 20 20",fill:"none",stroke:"currentColor",strokeWidth:"2",children:e?y.jsx("path",{d:"M6 6l8 8M14 6l-8 8"}):y.jsx("path",{d:"M3 5h14M3 10h14M3 15h14"})})}),e&&y.jsx("span",{className:"text-cistern-accent font-mono font-bold text-lg whitespace-nowrap",children:"Cistern"})]}),y.jsx("nav",{className:"flex-1 py-2 space-y-1 overflow-y-auto",children:Mg.map(r=>y.jsxs(jg,{to:r.to,className:()=>`flex items-center gap-3 px-4 py-2.5 mx-2 rounded-md transition-colors text-sm whitespace-nowrap ${n.pathname===r.to||r.to==="/app/"&&n.pathname==="/app"?"bg-cistern-accent/10 text-cistern-accent":"text-cistern-muted hover:text-cistern-fg hover:bg-cistern-border/30"}`,children:[y.jsx(r.icon,{className:"w-5 h-5 shrink-0"}),e&&y.jsx("span",{children:r.label})]},r.to))})]})]})}function Og({className:e}){return y.jsxs("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:[y.jsx("rect",{x:"3",y:"3",width:"7",height:"7"}),y.jsx("rect",{x:"14",y:"3",width:"7",height:"7"}),y.jsx("rect",{x:"14",y:"14",width:"7",height:"7"}),y.jsx("rect",{x:"3",y:"14",width:"7",height:"7"})]})}function Fg({className:e}){return y.jsx("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:y.jsx("path",{d:"M12 2.69l5.66 5.66a8 8 0 1 1-11.31 0z"})})}function Ig({className:e}){return y.jsxs("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:[y.jsx("circle",{cx:"12",cy:"12",r:"3"}),y.jsx("path",{d:"M12 1v4M12 19v4M4.22 4.22l2.83 2.83M16.95 16.95l2.83 2.83M1 12h4M19 12h4M4.22 19.78l2.83-2.83M16.95 7.05l2.83-2.83"})]})}function Ug({className:e}){return y.jsx("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:y.jsx("path",{d:"M22 12h-4l-3 9L9 3l-3 9H2"})})}function Ag({className:e}){return y.jsxs("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:[y.jsx("path",{d:"M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z"}),y.jsx("polyline",{points:"14 2 14 8 20 8"}),y.jsx("line",{x1:"16",y1:"13",x2:"8",y2:"13"}),y.jsx("line",{x1:"16",y1:"17",x2:"8",y2:"17"}),y.jsx("polyline",{points:"10 9 9 9 8 9"})]})}function Bg({className:e}){return y.jsx("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:y.jsx("path",{d:"M9 19c-5 1.5-5-2.5-7-3m14 6v-3.87a3.37 3.37 0 0 0-.94-2.61c3.14-.35 6.44-1.54 6.44-7A5.44 5.44 0 0 0 20 4.77 5.07 5.07 0 0 0 19.91 1S18.73.65 16 2.48a13.38 13.38 0 0 0-7 0C6.27.65 5.09 1 5.09 1A5.07 5.07 0 0 0 5 4.77a5.44 5.44 0 0 0-1.5 3.35c0 5.46 3.3 6.65 6.44 7A3.37 3.37 0 0 0 9 18.13V22"})})}function $g({data:e,connected:t,onMenuClick:n}){return y.jsxs("header",{className:"h-[60px] bg-cistern-surface border-b border-cistern-border flex items-center px-4 gap-4 shrink-0",children:[y.jsx("button",{onClick:n,className:"md:hidden text-cistern-muted hover:text-cistern-fg transition-colors","aria-label":"Open menu",children:y.jsx("svg",{width:"20",height:"20",viewBox:"0 0 20 20",fill:"none",stroke:"currentColor",strokeWidth:"2",children:y.jsx("path",{d:"M3 5h14M3 10h14M3 15h14"})})}),y.jsx("div",{className:"hidden md:flex items-center gap-2",children:y.jsx("span",{className:"text-cistern-accent font-mono font-bold text-lg",children:"Cistern"})}),y.jsx("div",{className:"flex-1"}),e&&y.jsxs("div",{className:"flex items-center gap-4 text-sm font-mono",children:[y.jsx(Lo,{label:"Flowing",count:e.flowing_count,color:"green"}),y.jsx(Lo,{label:"Queued",count:e.queued_count,color:"yellow"}),y.jsx(Lo,{label:"Delivered",count:e.done_count,color:"blue"})]}),y.jsxs("div",{className:"flex items-center gap-2",children:[y.jsx("div",{className:`w-2 h-2 rounded-full ${t?"bg-cistern-green":"bg-cistern-red"}`}),y.jsx("span",{className:"text-xs text-cistern-muted hidden sm:inline",children:t?"Live":"Disconnected"})]})]})}function Lo({label:e,count:t,color:n}){const r={green:"bg-cistern-green/20 text-cistern-green",yellow:"bg-cistern-yellow/20 text-cistern-yellow",blue:"bg-cistern-accent/20 text-cistern-accent"};return y.jsxs("span",{className:`inline-flex items-center gap-1.5 px-2 py-0.5 rounded-full text-xs ${r[n]}`,children:[e," ",t]})}function Vg(e={}){const{onData:t,onError:n,enabled:r=!0}=e,[l,i]=P.useState(null),[o,a]=P.useState(!1),[u,s]=P.useState(null),d=P.useRef(null),f=P.useRef(),p=P.useRef(!0),x=P.useRef(t),E=P.useRef(n),k=P.useRef(r);P.useEffect(()=>{x.current=t},[t]),P.useEffect(()=>{E.current=n},[n]),P.useEffect(()=>{k.current=r},[r]);const T=P.useCallback(()=>{d.current&&d.current.close();const h=new EventSource("/api/dashboard/events");h.onopen=()=>{p.current&&(a(!0),s(null))},h.onmessage=c=>{var v;if(p.current)try{const C=JSON.parse(c.data);i(C),(v=x.current)==null||v.call(x,C)}catch(C){console.warn("Failed to parse SSE message:",C,c.data)}},h.onerror=()=>{var v;if(!p.current)return;a(!1);const c=new Error("SSE connection lost");s(c),(v=E.current)==null||v.call(E,c),h.close(),d.current=null,k.current&&(clearTimeout(f.current),f.current=setTimeout(()=>{p.current&&T()},3e3))},d.current=h},[]);return P.useEffect(()=>{if(p.current=!0,!r){d.current&&(d.current.close(),d.current=null);return}return T(),()=>{p.current=!1,clearTimeout(f.current),d.current&&(d.current.close(),d.current=null)}},[T,r]),{data:l,connected:o,error:u}}const bd=P.createContext({data:null,connected:!1,error:null});function Hg({children:e}){const{data:t,connected:n,error:r}=Vg();return y.jsx(bd.Provider,{value:{data:t,connected:n,error:r},children:e})}function ep(){return P.useContext(bd)}function Wg(){return y.jsx(Hg,{children:y.jsx(Qg,{})})}function Qg(){const[e,t]=P.useState(!0),{data:n,connected:r}=ep();return y.jsxs("div",{className:"h-screen flex overflow-hidden bg-cistern-bg",children:[y.jsx(zg,{open:e,onToggle:()=>t(!e)}),y.jsxs("div",{className:"flex-1 flex flex-col min-w-0 overflow-hidden",children:[y.jsx($g,{data:n,connected:r,onMenuClick:()=>t(!e)}),y.jsx("main",{className:"flex-1 overflow-y-auto",children:y.jsx(ag,{})})]})]})}const Kg={open:{bg:"bg-cistern-yellow/20",text:"text-cistern-yellow",label:"Open"},in_progress:{bg:"bg-cistern-accent/20",text:"text-cistern-accent",label:"In Progress"},done:{bg:"bg-cistern-green/20",text:"text-cistern-green",label:"Done"},pooled:{bg:"bg-cistern-red/20",text:"text-cistern-red",label:"Pooled"},closed:{bg:"bg-cistern-muted/20",text:"text-cistern-muted",label:"Closed"}};function Yg({status:e,size:t="sm"}){const n=Kg[e]||{bg:"bg-cistern-muted/20",text:"text-cistern-muted",label:e},r=t==="sm"?"text-xs px-1.5 py-0.5":"text-sm px-2 py-1";return y.jsx("span",{className:`inline-flex items-center rounded-full font-mono font-medium ${n.bg} ${n.text} ${r}`,children:n.label})}function To(e){const t=Math.floor(e/1e9),n=Math.floor(t/3600),r=Math.floor(t%3600/60),l=t%60;return n>0?`${n}h${r.toString().padStart(2,"0")}m`:`${r}:${l.toString().padStart(2,"0")}`}function wc({cataractae:e,activity:t,isFlowing:n,onPeek:r}){const[l,i]=P.useState(To(e.elapsed)),o=P.useRef(Date.now()-e.elapsed/1e6);P.useEffect(()=>{o.current=Date.now()-e.elapsed/1e6,i(To(e.elapsed))},[e.elapsed]),P.useEffect(()=>{if(!n)return;const u=setInterval(()=>{const s=Date.now()-o.current;i(To(s*1e6))},1e3);return()=>clearInterval(u)},[n]);const a=e.total_cataractae>0?e.cataractae_index/e.total_cataractae*100:0;return y.jsxs("div",{className:`group border rounded-lg p-4 transition-all cursor-pointer ${n?"border-cistern-accent/40 bg-cistern-surface/80 hover:border-cistern-accent pulse-glow":"border-cistern-border bg-cistern-surface/40 opacity-60 hover:opacity-80"}`,onClick:()=>r(e.name),children:[y.jsxs("div",{className:"flex items-center justify-between mb-3",children:[y.jsxs("div",{className:"flex items-center gap-2",children:[y.jsx("span",{className:"font-mono font-bold text-cistern-fg",children:e.name}),y.jsx("span",{className:"text-xs text-cistern-muted",children:e.repo_name})]}),y.jsxs("div",{className:"flex items-center gap-2",children:[n&&y.jsx("span",{className:"text-xs font-mono text-cistern-accent",children:l}),y.jsx("button",{onClick:u=>{u.stopPropagation(),r(e.name)},className:"text-xs px-2 py-1 rounded border border-cistern-border text-cistern-muted hover:text-cistern-fg hover:border-cistern-muted transition-colors",children:"Peek"})]})]}),e.droplet_id&&n&&y.jsxs("div",{className:"mb-3 text-sm",children:[y.jsx("span",{className:"font-mono text-cistern-green",children:e.droplet_id}),y.jsx("span",{className:"text-cistern-muted mx-2",children:"·"}),y.jsx("span",{className:"text-cistern-fg",children:e.title})]}),y.jsx(Xg,{steps:e.steps,currentIndex:e.cataractae_index-1,isFlowing:n}),a>0&&y.jsx("div",{className:"mt-3",children:y.jsx("div",{className:"h-1.5 bg-cistern-border rounded-full overflow-hidden",children:y.jsx("div",{className:`h-full rounded-full transition-all duration-500 ${n?"bg-cistern-accent":"bg-cistern-muted"}`,style:{width:`${a}%`}})})}),t&&t.recent_notes.length>0&&y.jsx("div",{className:"mt-3 space-y-1",children:t.recent_notes.slice(0,3).map(u=>y.jsx(Gg,{note:u},u.id))})]})}function Xg({steps:e,currentIndex:t,isFlowing:n}){return y.jsx("div",{className:"flex items-center gap-0 overflow-x-auto pb-1",children:e.map((r,l)=>{const i=l===t&&n,o=l<t,a=l>t;return y.jsxs("div",{className:"flex items-center shrink-0",children:[l>0&&y.jsx("div",{className:`h-0.5 w-3 ${o?"bg-cistern-green":i?"bg-cistern-accent":"bg-cistern-border"}`}),y.jsxs("div",{className:`relative px-2 py-1 rounded text-xs font-mono whitespace-nowrap ${i?"water-flow-active text-cistern-bg font-bold":o?"bg-cistern-green/20 text-cistern-green":a&&n?"bg-cistern-accent/10 text-cistern-accent/50":"bg-cistern-border/30 text-cistern-muted"}`,children:[r,i&&y.jsx("span",{className:"absolute -top-1 -right-1 w-2 h-2 bg-cistern-accent rounded-full animate-ping"})]})]},r)})})}function Gg({note:e}){return y.jsxs("div",{className:"text-xs text-cistern-muted pl-2 border-l border-cistern-border",children:[y.jsx("span",{className:"text-cistern-accent font-mono",children:e.cataractae_name}),y.jsx("span",{className:"mx-1",children:"·"}),y.jsx("span",{children:e.content.length>100?e.content.slice(0,100)+"…":e.content})]})}function Gi({droplet:e,blockedBy:t}){const n=Jg(e.created_at);return y.jsxs("div",{className:"w-full flex items-center gap-3 px-3 py-2 rounded-md hover:bg-cistern-border/20 transition-colors text-left",children:[y.jsx(Yg,{status:e.status}),y.jsx("span",{className:"font-mono text-xs text-cistern-accent",children:e.id}),y.jsx("span",{className:"text-sm text-cistern-fg truncate flex-1",children:e.title}),t&&y.jsxs("span",{className:"text-xs text-cistern-yellow",title:`Blocked by ${t}`,children:["⛏ ",t]}),y.jsx("span",{className:"text-xs text-cistern-muted whitespace-nowrap",children:n})]})}function Jg(e){const t=new Date(e).getTime(),r=Date.now()-t,l=Math.floor(r/6e4);if(l<60)return`${l}m`;const i=Math.floor(l/60);return i<24?`${i}h`:`${Math.floor(i/24)}d`}const xc=50*1024;function Zg({aqueductName:e,onClose:t}){const[n,r]=P.useState(""),[l,i]=P.useState(!1),o=P.useRef(null),a=P.useRef(null),u=P.useRef(!0),s=P.useCallback(d=>{r(f=>{const p=f+d;return p.length>xc?p.slice(p.length-xc):p})},[]);return P.useEffect(()=>{u.current=!0;const f=`${window.location.protocol==="https:"?"wss:":"ws:"}//${window.location.host}/ws/aqueducts/${encodeURIComponent(e)}/peek`,p=new WebSocket(f);return a.current=p,p.onopen=()=>{u.current&&i(!0)},p.onmessage=x=>{u.current&&s(x.data)},p.onclose=()=>{u.current&&i(!1)},p.onerror=()=>{u.current&&i(!1)},()=>{u.current=!1,p.close(),a.current=null}},[e,s]),P.useEffect(()=>{o.current&&(o.current.scrollTop=o.current.scrollHeight)},[n]),y.jsxs("div",{className:"fixed inset-y-0 right-0 w-full md:w-[600px] bg-cistern-bg border-l border-cistern-border shadow-2xl z-50 flex flex-col",children:[y.jsxs("div",{className:"flex items-center justify-between px-4 py-3 border-b border-cistern-border",children:[y.jsxs("div",{className:"flex items-center gap-3",children:[y.jsx("h3",{className:"font-mono text-cistern-accent",children:e}),y.jsx("span",{className:"text-xs text-cistern-muted",children:"Peek"}),y.jsx("div",{className:`w-2 h-2 rounded-full ${l?"bg-cistern-green":"bg-cistern-red"}`})]}),y.jsx("button",{onClick:t,className:"text-cistern-muted hover:text-cistern-fg transition-colors text-lg leading-none",children:"×"})]}),y.jsx("pre",{ref:o,className:"flex-1 overflow-auto p-4 font-mono text-xs text-cistern-green bg-cistern-bg whitespace-pre-wrap break-all",children:n||"Connecting…"})]})}function qg(){const{data:e,error:t}=ep(),[n,r]=P.useState(null),l=P.useMemo(()=>{const o={};if(e)for(const a of e.flow_activities)o[a.droplet_id]=a;return o},[e]);if(t&&!e)return y.jsx("div",{className:"flex items-center justify-center h-full",children:y.jsxs("div",{className:"text-center",children:[y.jsx("div",{className:"text-cistern-red text-lg font-mono mb-2",children:"Connection Error"}),y.jsx("div",{className:"text-cistern-muted text-sm",children:t.message})]})});if(!e)return y.jsx("div",{className:"flex items-center justify-center h-full",children:y.jsx("div",{className:"text-cistern-muted font-mono",children:"Loading dashboard…"})});const i=new Set(e.cataractae.filter(o=>o.droplet_id).map(o=>o.droplet_id));return y.jsxs("div",{className:"flex-1 overflow-y-auto p-4 md:p-6 space-y-6",children:[y.jsx(bg,{cataractae:e.cataractae,flowingIds:i,activityMap:l,onPeek:r}),y.jsx(ey,{data:e}),n&&y.jsx(Zg,{aqueductName:n,onClose:()=>r(null)})]})}function bg({cataractae:e,flowingIds:t,activityMap:n,onPeek:r}){const l=e.filter(o=>o.droplet_id&&t.has(o.droplet_id)),i=e.filter(o=>!o.droplet_id);return y.jsxs("section",{children:[y.jsx("h2",{className:"text-sm font-mono text-cistern-muted uppercase tracking-wider mb-3",children:"Aqueducts"}),l.length>0&&y.jsx("div",{className:"space-y-3 mb-4",children:l.map(o=>y.jsx(wc,{cataractae:o,activity:n[o.droplet_id||""],isFlowing:!0,onPeek:r},o.name))}),i.length>0&&y.jsx("div",{className:"space-y-2",children:i.map(o=>y.jsx(wc,{cataractae:o,isFlowing:!1,onPeek:r},o.name))})]})}function ey({data:e}){return y.jsxs("div",{className:"grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4",children:[y.jsx(ty,{data:e}),y.jsx(ny,{items:e.cistern_items,blockedByMap:e.blocked_by_map}),y.jsx(ry,{items:e.pooled_items}),y.jsx(ly,{items:e.unassigned_items}),y.jsx(iy,{items:e.recent_items})]})}function ty({data:e}){return y.jsxs("div",{className:"bg-cistern-surface border border-cistern-border rounded-lg p-4",children:[y.jsx("h3",{className:"text-sm font-mono text-cistern-muted uppercase tracking-wider mb-3",children:"Cistern"}),y.jsxs("div",{className:"space-y-2",children:[y.jsxs("div",{className:"flex items-center justify-between",children:[y.jsx("span",{className:"text-sm text-cistern-fg",children:"Flowing"}),y.jsx("span",{className:"font-mono text-cistern-green",children:e.flowing_count})]}),y.jsxs("div",{className:"flex items-center justify-between",children:[y.jsx("span",{className:"text-sm text-cistern-fg",children:"Queued"}),y.jsx("span",{className:"font-mono text-cistern-yellow",children:e.queued_count})]}),y.jsxs("div",{className:"flex items-center justify-between",children:[y.jsx("span",{className:"text-sm text-cistern-fg",children:"Delivered"}),y.jsx("span",{className:"font-mono text-cistern-accent",children:e.done_count})]}),y.jsxs("div",{className:"flex items-center justify-between",children:[y.jsx("span",{className:"text-sm text-cistern-fg",children:"Pooled"}),y.jsx("span",{className:"font-mono text-cistern-red",children:e.pooled_items.length})]})]}),y.jsx("div",{className:"mt-3 pt-3 border-t border-cistern-border",children:y.jsxs("div",{className:"flex items-center gap-2",children:[y.jsx("div",{className:`w-2 h-2 rounded-full ${e.farm_running?"bg-cistern-green":"bg-cistern-muted"}`}),y.jsxs("span",{className:"text-xs text-cistern-muted",children:["Castellarius ",e.farm_running?"Running":"Stopped"]})]})})]})}function ny({items:e,blockedByMap:t}){return e.length===0?null:y.jsxs("div",{className:"bg-cistern-surface border border-cistern-border rounded-lg p-4",children:[y.jsxs("h3",{className:"text-sm font-mono text-cistern-muted uppercase tracking-wider mb-3",children:["Queue (",e.length,")"]}),y.jsx("div",{className:"space-y-1 max-h-64 overflow-y-auto",children:e.map(n=>y.jsx(Gi,{droplet:n,blockedBy:t[n.id]},n.id))})]})}function ry({items:e}){return e.length===0?null:y.jsxs("div",{className:"bg-cistern-surface border border-cistern-border rounded-lg p-4",children:[y.jsxs("h3",{className:"text-sm font-mono text-cistern-red uppercase tracking-wider mb-3",children:["Pooled (",e.length,")"]}),y.jsx("div",{className:"space-y-1 max-h-64 overflow-y-auto",children:e.map(t=>y.jsx(Gi,{droplet:t},t.id))})]})}function ly({items:e}){return e.length===0?null:y.jsxs("div",{className:"bg-cistern-surface border border-cistern-yellow/30 rounded-lg p-4",children:[y.jsxs("h3",{className:"text-sm font-mono text-cistern-yellow uppercase tracking-wider mb-3 flex items-center gap-2",children:[y.jsx("span",{children:"⚠"})," Orphaned (",e.length,")"]}),y.jsx("div",{className:"space-y-1 max-h-64 overflow-y-auto",children:e.map(t=>y.jsx(Gi,{droplet:t},t.id))})]})}function iy({items:e}){return e.length===0?null:y.jsxs("div",{className:"bg-cistern-surface border border-cistern-border rounded-lg p-4",children:[y.jsx("h3",{className:"text-sm font-mono text-cistern-muted uppercase tracking-wider mb-3",children:"Recent Flow"}),y.jsx("div",{className:"space-y-1 max-h-64 overflow-y-auto",children:e.slice(0,5).map(t=>y.jsx(Gi,{droplet:t},t.id))})]})}function Mr({title:e}){return y.jsx("div",{className:"flex items-center justify-center h-full",children:y.jsxs("div",{className:"text-center",children:[y.jsx("h2",{className:"text-lg font-mono text-cistern-fg mb-2",children:e}),y.jsx("p",{className:"text-sm text-cistern-muted",children:"Coming soon"})]})})}const oy=mg([{path:"/app",element:y.jsx(Wg,{}),children:[{index:!0,element:y.jsx(qg,{})},{path:"droplets",element:y.jsx(Mr,{title:"Droplets"})},{path:"castellarius",element:y.jsx(Mr,{title:"Castellarius"})},{path:"doctor",element:y.jsx(Mr,{title:"Doctor"})},{path:"logs",element:y.jsx(Mr,{title:"Logs"})},{path:"repos",element:y.jsx(Mr,{title:"Repos / Skills"})}]}]);Do.createRoot(document.getElementById("root")).render(y.jsx(Mc.StrictMode,{children:y.jsx(kg,{router:oy})}));
+ */function pr(){return pr=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},pr.apply(this,arguments)}function Zd(e,t){if(e==null)return{};var n={},r=Object.keys(e),l,i;for(i=0;i<r.length;i++)l=r[i],!(t.indexOf(l)>=0)&&(n[l]=e[l]);return n}function cg(e){return!!(e.metaKey||e.altKey||e.ctrlKey||e.shiftKey)}function fg(e,t){return e.button===0&&(!t||t==="_self")&&!cg(e)}const dg=["onClick","relative","reloadDocument","replace","state","target","to","preventScrollReset","viewTransition"],pg=["aria-current","caseSensitive","className","end","style","to","viewTransition","children"],hg="6";try{window.__reactRouterVersion=hg}catch{}function mg(e,t){return kv({basename:void 0,future:pr({},void 0,{v7_prependBasename:!0}),history:Km({window:void 0}),hydrationData:vg(),routes:e,mapRouteProperties:sg,dataStrategy:void 0,patchRoutesOnNavigation:void 0,window:void 0}).initialize()}function vg(){var e;let t=(e=window)==null?void 0:e.__staticRouterHydrationData;return t&&t.errors&&(t=pr({},t,{errors:gg(t.errors)})),t}function gg(e){if(!e)return null;let t=Object.entries(e),n={};for(let[r,l]of t)if(l&&l.__type==="RouteErrorResponse")n[r]=new Li(l.status,l.statusText,l.data,l.internal===!0);else if(l&&l.__type==="Error"){if(l.__subType){let i=window[l.__subType];if(typeof i=="function")try{let o=new i(l.message);o.stack="",n[r]=o}catch{}}if(n[r]==null){let i=new Error(l.message);i.stack="",n[r]=i}}else n[r]=l;return n}const qd=P.createContext({isTransitioning:!1}),yg=P.createContext(new Map),wg="startTransition",vc=Mp[wg],xg="flushSync",gc=Qm[xg];function Sg(e){vc?vc(e):e()}function Dr(e){gc?gc(e):e()}class Eg{constructor(){this.status="pending",this.promise=new Promise((t,n)=>{this.resolve=r=>{this.status==="pending"&&(this.status="resolved",t(r))},this.reject=r=>{this.status==="pending"&&(this.status="rejected",n(r))}})}}function kg(e){let{fallbackElement:t,router:n,future:r}=e,[l,i]=P.useState(n.state),[o,a]=P.useState(),[u,s]=P.useState({isTransitioning:!1}),[d,f]=P.useState(),[p,x]=P.useState(),[E,k]=P.useState(),T=P.useRef(new Map),{v7_startTransition:h}=r||{},c=P.useCallback(_=>{h?Sg(_):_()},[h]),v=P.useCallback((_,$)=>{let{deletedFetchers:D,flushSync:q,viewTransitionOpts:re}=$;_.fetchers.forEach((Pe,ut)=>{Pe.data!==void 0&&T.current.set(ut,Pe.data)}),D.forEach(Pe=>T.current.delete(Pe));let xe=n.window==null||n.window.document==null||typeof n.window.document.startViewTransition!="function";if(!re||xe){q?Dr(()=>i(_)):c(()=>i(_));return}if(q){Dr(()=>{p&&(d&&d.resolve(),p.skipTransition()),s({isTransitioning:!0,flushSync:!0,currentLocation:re.currentLocation,nextLocation:re.nextLocation})});let Pe=n.window.document.startViewTransition(()=>{Dr(()=>i(_))});Pe.finished.finally(()=>{Dr(()=>{f(void 0),x(void 0),a(void 0),s({isTransitioning:!1})})}),Dr(()=>x(Pe));return}p?(d&&d.resolve(),p.skipTransition(),k({state:_,currentLocation:re.currentLocation,nextLocation:re.nextLocation})):(a(_),s({isTransitioning:!0,flushSync:!1,currentLocation:re.currentLocation,nextLocation:re.nextLocation}))},[n.window,p,d,T,c]);P.useLayoutEffect(()=>n.subscribe(v),[n,v]),P.useEffect(()=>{u.isTransitioning&&!u.flushSync&&f(new Eg)},[u]),P.useEffect(()=>{if(d&&o&&n.window){let _=o,$=d.promise,D=n.window.document.startViewTransition(async()=>{c(()=>i(_)),await $});D.finished.finally(()=>{f(void 0),x(void 0),a(void 0),s({isTransitioning:!1})}),x(D)}},[c,o,d,n.window]),P.useEffect(()=>{d&&o&&l.location.key===o.location.key&&d.resolve()},[d,p,l.location,o]),P.useEffect(()=>{!u.isTransitioning&&E&&(a(E.state),s({isTransitioning:!0,flushSync:!1,currentLocation:E.currentLocation,nextLocation:E.nextLocation}),k(void 0))},[u.isTransitioning,E]),P.useEffect(()=>{},[]);let C=P.useMemo(()=>({createHref:n.createHref,encodeLocation:n.encodeLocation,go:_=>n.navigate(_),push:(_,$,D)=>n.navigate(_,{state:$,preventScrollReset:D==null?void 0:D.preventScrollReset}),replace:(_,$,D)=>n.navigate(_,{replace:!0,state:$,preventScrollReset:D==null?void 0:D.preventScrollReset})}),[n]),j=n.basename||"/",z=P.useMemo(()=>({router:n,navigator:C,static:!1,basename:j}),[n,C,j]),g=P.useMemo(()=>({v7_relativeSplatPath:n.future.v7_relativeSplatPath}),[n.future.v7_relativeSplatPath]);return P.useEffect(()=>og(r,n.future),[r,n.future]),P.createElement(P.Fragment,null,P.createElement(yl.Provider,{value:z},P.createElement(ju.Provider,{value:l},P.createElement(yg.Provider,{value:T.current},P.createElement(qd.Provider,{value:u},P.createElement(ug,{basename:j,location:l.location,navigationType:l.historyAction,navigator:C,future:g},l.initialized||n.future.v7_partialHydration?P.createElement(Cg,{routes:n.routes,future:n.future,state:l}):t))))),null)}const Cg=P.memo(Pg);function Pg(e){let{routes:t,future:n,state:r}=e;return Xv(t,void 0,r,n)}const _g=typeof window<"u"&&typeof window.document<"u"&&typeof window.document.createElement<"u",Ng=/^(?:[a-z][a-z0-9+.-]*:|\/\/)/i,Rg=P.forwardRef(function(t,n){let{onClick:r,relative:l,reloadDocument:i,replace:o,state:a,target:u,to:s,preventScrollReset:d,viewTransition:f}=t,p=Zd(t,dg),{basename:x}=P.useContext(dn),E,k=!1;if(typeof s=="string"&&Ng.test(s)&&(E=s,_g))try{let v=new URL(window.location.href),C=s.startsWith("//")?new URL(v.protocol+s):new URL(s),j=Ft(C.pathname,x);C.origin===v.origin&&j!=null?s=j+C.search+C.hash:k=!0}catch{}let T=Hv(s,{relative:l}),h=Tg(s,{replace:o,state:a,target:u,preventScrollReset:d,relative:l,viewTransition:f});function c(v){r&&r(v),v.defaultPrevented||h(v)}return P.createElement("a",pr({},p,{href:E||T,onClick:k||i?r:c,ref:n,target:u}))}),jg=P.forwardRef(function(t,n){let{"aria-current":r="page",caseSensitive:l=!1,className:i="",end:o=!1,style:a,to:u,viewTransition:s,children:d}=t,f=Zd(t,pg),p=Xi(u,{relative:f.relative}),x=gr(),E=P.useContext(ju),{navigator:k,basename:T}=P.useContext(dn),h=E!=null&&Dg(p)&&s===!0,c=k.encodeLocation?k.encodeLocation(p).pathname:p.pathname,v=x.pathname,C=E&&E.navigation&&E.navigation.location?E.navigation.location.pathname:null;l||(v=v.toLowerCase(),C=C?C.toLowerCase():null,c=c.toLowerCase()),C&&T&&(C=Ft(C,T)||C);const j=c!=="/"&&c.endsWith("/")?c.length-1:c.length;let z=v===c||!o&&v.startsWith(c)&&v.charAt(j)==="/",g=C!=null&&(C===c||!o&&C.startsWith(c)&&C.charAt(c.length)==="/"),_={isActive:z,isPending:g,isTransitioning:h},$=z?r:void 0,D;typeof i=="function"?D=i(_):D=[i,z?"active":null,g?"pending":null,h?"transitioning":null].filter(Boolean).join(" ");let q=typeof a=="function"?a(_):a;return P.createElement(Rg,pr({},f,{"aria-current":$,className:D,ref:n,style:q,to:u,viewTransition:s}),typeof d=="function"?d(_):d)});var Ra;(function(e){e.UseScrollRestoration="useScrollRestoration",e.UseSubmit="useSubmit",e.UseSubmitFetcher="useSubmitFetcher",e.UseFetcher="useFetcher",e.useViewTransitionState="useViewTransitionState"})(Ra||(Ra={}));var yc;(function(e){e.UseFetcher="useFetcher",e.UseFetchers="useFetchers",e.UseScrollRestoration="useScrollRestoration"})(yc||(yc={}));function Lg(e){let t=P.useContext(yl);return t||K(!1),t}function Tg(e,t){let{target:n,replace:r,state:l,preventScrollReset:i,relative:o,viewTransition:a}=t===void 0?{}:t,u=Wv(),s=gr(),d=Xi(e,{relative:o});return P.useCallback(f=>{if(fg(f,n)){f.preventDefault();let p=r!==void 0?r:Mn(s)===Mn(d);u(e,{replace:p,state:l,preventScrollReset:i,relative:o,viewTransition:a})}},[s,u,d,r,l,n,e,i,o,a])}function Dg(e,t){t===void 0&&(t={});let n=P.useContext(qd);n==null&&K(!1);let{basename:r}=Lg(Ra.useViewTransitionState),l=Xi(e,{relative:t.relative});if(!n.isTransitioning)return!1;let i=Ft(n.currentLocation.pathname,r)||n.currentLocation.pathname,o=Ft(n.nextLocation.pathname,r)||n.nextLocation.pathname;return ji(l.pathname,o)!=null||ji(l.pathname,i)!=null}const Mg=[{to:"/app/",label:"Dashboard",icon:Og},{to:"/app/droplets",label:"Droplets",icon:Fg},{to:"/app/castellarius",label:"Castellarius",icon:Ig},{to:"/app/doctor",label:"Doctor",icon:Ug},{to:"/app/logs",label:"Logs",icon:Ag},{to:"/app/repos",label:"Repos/Skills",icon:Bg}];function zg({open:e,onToggle:t}){const n=gr();return y.jsxs(y.Fragment,{children:[e&&y.jsx("div",{className:"fixed inset-0 bg-black/50 z-20 md:hidden",onClick:t}),y.jsxs("aside",{className:`fixed md:relative z-30 h-full bg-cistern-surface border-r border-cistern-border flex flex-col transition-all duration-200 ${e?"w-56":"w-0 md:w-14"} overflow-hidden`,children:[y.jsxs("div",{className:"flex items-center gap-3 px-4 py-4 border-b border-cistern-border min-h-[60px]",children:[y.jsx("button",{onClick:t,className:"text-cistern-muted hover:text-cistern-fg transition-colors","aria-label":"Toggle sidebar",children:y.jsx("svg",{width:"20",height:"20",viewBox:"0 0 20 20",fill:"none",stroke:"currentColor",strokeWidth:"2",children:e?y.jsx("path",{d:"M6 6l8 8M14 6l-8 8"}):y.jsx("path",{d:"M3 5h14M3 10h14M3 15h14"})})}),e&&y.jsx("span",{className:"text-cistern-accent font-mono font-bold text-lg whitespace-nowrap",children:"Cistern"})]}),y.jsx("nav",{className:"flex-1 py-2 space-y-1 overflow-y-auto",children:Mg.map(r=>y.jsxs(jg,{to:r.to,className:()=>`flex items-center gap-3 px-4 py-2.5 mx-2 rounded-md transition-colors text-sm whitespace-nowrap ${n.pathname===r.to||r.to==="/app/"&&n.pathname==="/app"?"bg-cistern-accent/10 text-cistern-accent":"text-cistern-muted hover:text-cistern-fg hover:bg-cistern-border/30"}`,children:[y.jsx(r.icon,{className:"w-5 h-5 shrink-0"}),e&&y.jsx("span",{children:r.label})]},r.to))})]})]})}function Og({className:e}){return y.jsxs("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:[y.jsx("rect",{x:"3",y:"3",width:"7",height:"7"}),y.jsx("rect",{x:"14",y:"3",width:"7",height:"7"}),y.jsx("rect",{x:"14",y:"14",width:"7",height:"7"}),y.jsx("rect",{x:"3",y:"14",width:"7",height:"7"})]})}function Fg({className:e}){return y.jsx("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:y.jsx("path",{d:"M12 2.69l5.66 5.66a8 8 0 1 1-11.31 0z"})})}function Ig({className:e}){return y.jsxs("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:[y.jsx("circle",{cx:"12",cy:"12",r:"3"}),y.jsx("path",{d:"M12 1v4M12 19v4M4.22 4.22l2.83 2.83M16.95 16.95l2.83 2.83M1 12h4M19 12h4M4.22 19.78l2.83-2.83M16.95 7.05l2.83-2.83"})]})}function Ug({className:e}){return y.jsx("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:y.jsx("path",{d:"M22 12h-4l-3 9L9 3l-3 9H2"})})}function Ag({className:e}){return y.jsxs("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:[y.jsx("path",{d:"M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z"}),y.jsx("polyline",{points:"14 2 14 8 20 8"}),y.jsx("line",{x1:"16",y1:"13",x2:"8",y2:"13"}),y.jsx("line",{x1:"16",y1:"17",x2:"8",y2:"17"}),y.jsx("polyline",{points:"10 9 9 9 8 9"})]})}function Bg({className:e}){return y.jsx("svg",{className:e,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:y.jsx("path",{d:"M9 19c-5 1.5-5-2.5-7-3m14 6v-3.87a3.37 3.37 0 0 0-.94-2.61c3.14-.35 6.44-1.54 6.44-7A5.44 5.44 0 0 0 20 4.77 5.07 5.07 0 0 0 19.91 1S18.73.65 16 2.48a13.38 13.38 0 0 0-7 0C6.27.65 5.09 1 5.09 1A5.07 5.07 0 0 0 5 4.77a5.44 5.44 0 0 0-1.5 3.35c0 5.46 3.3 6.65 6.44 7A3.37 3.37 0 0 0 9 18.13V22"})})}function $g({data:e,connected:t,onMenuClick:n}){return y.jsxs("header",{className:"h-[60px] bg-cistern-surface border-b border-cistern-border flex items-center px-4 gap-4 shrink-0",children:[y.jsx("button",{onClick:n,className:"md:hidden text-cistern-muted hover:text-cistern-fg transition-colors","aria-label":"Open menu",children:y.jsx("svg",{width:"20",height:"20",viewBox:"0 0 20 20",fill:"none",stroke:"currentColor",strokeWidth:"2",children:y.jsx("path",{d:"M3 5h14M3 10h14M3 15h14"})})}),y.jsx("div",{className:"hidden md:flex items-center gap-2",children:y.jsx("span",{className:"text-cistern-accent font-mono font-bold text-lg",children:"Cistern"})}),y.jsx("div",{className:"flex-1"}),e&&y.jsxs("div",{className:"flex items-center gap-4 text-sm font-mono",children:[y.jsx(Lo,{label:"Flowing",count:e.flowing_count,color:"green"}),y.jsx(Lo,{label:"Queued",count:e.queued_count,color:"yellow"}),y.jsx(Lo,{label:"Delivered",count:e.done_count,color:"blue"})]}),y.jsxs("div",{className:"flex items-center gap-2",children:[y.jsx("div",{className:`w-2 h-2 rounded-full ${t?"bg-cistern-green":"bg-cistern-red"}`}),y.jsx("span",{className:"text-xs text-cistern-muted hidden sm:inline",children:t?"Live":"Disconnected"})]})]})}function Lo({label:e,count:t,color:n}){const r={green:"bg-cistern-green/20 text-cistern-green",yellow:"bg-cistern-yellow/20 text-cistern-yellow",blue:"bg-cistern-accent/20 text-cistern-accent"};return y.jsxs("span",{className:`inline-flex items-center gap-1.5 px-2 py-0.5 rounded-full text-xs ${r[n]}`,children:[e," ",t]})}function Vg(e={}){const{onData:t,onError:n,enabled:r=!0}=e,[l,i]=P.useState(null),[o,a]=P.useState(!1),[u,s]=P.useState(null),d=P.useRef(null),f=P.useRef(),p=P.useRef(!0),x=P.useRef(t),E=P.useRef(n),k=P.useRef(r);P.useEffect(()=>{x.current=t},[t]),P.useEffect(()=>{E.current=n},[n]),P.useEffect(()=>{k.current=r},[r]);const T=P.useCallback(()=>{d.current&&d.current.close();const h=new EventSource("/api/dashboard/events");h.onopen=()=>{p.current&&(a(!0),s(null))},h.onmessage=c=>{var v;if(p.current)try{const C=JSON.parse(c.data);i(C),(v=x.current)==null||v.call(x,C)}catch(C){console.warn("Failed to parse SSE message:",C,c.data)}},h.onerror=()=>{var v;if(!p.current)return;a(!1);const c=new Error("SSE connection lost");s(c),(v=E.current)==null||v.call(E,c),h.close(),d.current=null,k.current&&(clearTimeout(f.current),f.current=setTimeout(()=>{p.current&&T()},3e3))},d.current=h},[]);return P.useEffect(()=>{if(p.current=!0,!r){d.current&&(d.current.close(),d.current=null);return}return T(),()=>{p.current=!1,clearTimeout(f.current),d.current&&(d.current.close(),d.current=null)}},[T,r]),{data:l,connected:o,error:u}}const bd=P.createContext({data:null,connected:!1,error:null});function Hg({children:e}){const{data:t,connected:n,error:r}=Vg();return y.jsx(bd.Provider,{value:{data:t,connected:n,error:r},children:e})}function ep(){return P.useContext(bd)}function Wg(){return y.jsx(Hg,{children:y.jsx(Qg,{})})}function Qg(){const[e,t]=P.useState(!0),{data:n,connected:r}=ep();return y.jsxs("div",{className:"h-screen flex overflow-hidden bg-cistern-bg",children:[y.jsx(zg,{open:e,onToggle:()=>t(!e)}),y.jsxs("div",{className:"flex-1 flex flex-col min-w-0 overflow-hidden",children:[y.jsx($g,{data:n,connected:r,onMenuClick:()=>t(!e)}),y.jsx("main",{className:"flex-1 overflow-y-auto",children:y.jsx(ag,{})})]})]})}const Kg={open:{bg:"bg-cistern-yellow/20",text:"text-cistern-yellow",label:"Open"},in_progress:{bg:"bg-cistern-accent/20",text:"text-cistern-accent",label:"In Progress"},done:{bg:"bg-cistern-green/20",text:"text-cistern-green",label:"Done"},pooled:{bg:"bg-cistern-red/20",text:"text-cistern-red",label:"Pooled"},closed:{bg:"bg-cistern-muted/20",text:"text-cistern-muted",label:"Closed"}};function Yg({status:e,size:t="sm"}){const n=Kg[e]||{bg:"bg-cistern-muted/20",text:"text-cistern-muted",label:e},r=t==="sm"?"text-xs px-1.5 py-0.5":"text-sm px-2 py-1";return y.jsx("span",{className:`inline-flex items-center rounded-full font-mono font-medium ${n.bg} ${n.text} ${r}`,children:n.label})}function To(e){const t=Math.floor(e/1e9),n=Math.floor(t/3600),r=Math.floor(t%3600/60),l=t%60;return n>0?`${n}h${r.toString().padStart(2,"0")}m`:`${r}:${l.toString().padStart(2,"0")}`}function wc({cataractae:e,activity:t,isFlowing:n,onPeek:r}){const[l,i]=P.useState(To(e.elapsed)),o=P.useRef(Date.now()-e.elapsed/1e6);P.useEffect(()=>{o.current=Date.now()-e.elapsed/1e6,i(To(e.elapsed))},[e.elapsed]),P.useEffect(()=>{if(!n)return;const u=setInterval(()=>{const s=Date.now()-o.current;i(To(s*1e6))},1e3);return()=>clearInterval(u)},[n]);const a=e.total_cataractae>0?e.cataractae_index/e.total_cataractae*100:0;return y.jsxs("div",{className:`group border rounded-lg p-4 transition-all cursor-pointer ${n?"border-cistern-accent/40 bg-cistern-surface/80 hover:border-cistern-accent pulse-glow":"border-cistern-border bg-cistern-surface/40 opacity-60 hover:opacity-80"}`,onClick:()=>r(e.name),children:[y.jsxs("div",{className:"flex items-center justify-between mb-3",children:[y.jsxs("div",{className:"flex items-center gap-2",children:[y.jsx("span",{className:"font-mono font-bold text-cistern-fg",children:e.name}),y.jsx("span",{className:"text-xs text-cistern-muted",children:e.repo_name})]}),y.jsxs("div",{className:"flex items-center gap-2",children:[n&&y.jsx("span",{className:"text-xs font-mono text-cistern-accent",children:l}),y.jsx("button",{onClick:u=>{u.stopPropagation(),r(e.name)},className:"text-xs px-2 py-1 rounded border border-cistern-border text-cistern-muted hover:text-cistern-fg hover:border-cistern-muted transition-colors",children:"Peek"})]})]}),e.droplet_id&&n&&y.jsxs("div",{className:"mb-3 text-sm",children:[y.jsx("span",{className:"font-mono text-cistern-green",children:e.droplet_id}),y.jsx("span",{className:"text-cistern-muted mx-2",children:"·"}),y.jsx("span",{className:"text-cistern-fg",children:e.title})]}),y.jsx(Xg,{steps:e.steps,currentIndex:e.cataractae_index-1,isFlowing:n}),a>0&&y.jsx("div",{className:"mt-3",children:y.jsx("div",{className:"h-1.5 bg-cistern-border rounded-full overflow-hidden",children:y.jsx("div",{className:`h-full rounded-full transition-all duration-500 ${n?"bg-cistern-accent":"bg-cistern-muted"}`,style:{width:`${a}%`}})})}),t&&t.recent_notes.length>0&&y.jsx("div",{className:"mt-3 space-y-1",children:t.recent_notes.slice(0,3).map(u=>y.jsx(Gg,{note:u},u.id))})]})}function Xg({steps:e,currentIndex:t,isFlowing:n}){return y.jsx("div",{className:"flex items-center gap-0 overflow-x-auto pb-1",children:e.map((r,l)=>{const i=l===t&&n,o=l<t,a=l>t;return y.jsxs("div",{className:"flex items-center shrink-0",children:[l>0&&y.jsx("div",{className:`h-0.5 w-3 ${o?"bg-cistern-green":i?"bg-cistern-accent":"bg-cistern-border"}`}),y.jsxs("div",{className:`relative px-2 py-1 rounded text-xs font-mono whitespace-nowrap ${i?"water-flow-active text-cistern-bg font-bold":o?"bg-cistern-green/20 text-cistern-green":a&&n?"bg-cistern-accent/10 text-cistern-accent/50":"bg-cistern-border/30 text-cistern-muted"}`,children:[r,i&&y.jsx("span",{className:"absolute -top-1 -right-1 w-2 h-2 bg-cistern-accent rounded-full animate-ping"})]})]},r)})})}function Gg({note:e}){return y.jsxs("div",{className:"text-xs text-cistern-muted pl-2 border-l border-cistern-border",children:[y.jsx("span",{className:"text-cistern-accent font-mono",children:e.cataractae_name}),y.jsx("span",{className:"mx-1",children:"·"}),y.jsx("span",{children:e.content.length>100?e.content.slice(0,100)+"…":e.content})]})}function Jg(e){const t=new Date(e).getTime();if(isNaN(t))return"--";const r=Date.now()-t,l=Math.floor(r/6e4);if(l<60)return`${l}m`;const i=Math.floor(l/60);return i<24?`${i}h`:`${Math.floor(i/24)}d`}function Gi({droplet:e,blockedBy:t}){const n=Jg(e.created_at);return y.jsxs("div",{className:"w-full flex items-center gap-3 px-3 py-2 rounded-md hover:bg-cistern-border/20 transition-colors text-left",children:[y.jsx(Yg,{status:e.status}),y.jsx("span",{className:"font-mono text-xs text-cistern-accent",children:e.id}),y.jsx("span",{className:"text-sm text-cistern-fg truncate flex-1",children:e.title}),t&&y.jsxs("span",{className:"text-xs text-cistern-yellow",title:`Blocked by ${t}`,children:["⛏ ",t]}),y.jsx("span",{className:"text-xs text-cistern-muted whitespace-nowrap",children:n})]})}const xc=50*1024;function Zg({aqueductName:e,onClose:t}){const[n,r]=P.useState(""),[l,i]=P.useState(!1),o=P.useRef(null),a=P.useRef(null),u=P.useRef(!0),s=P.useCallback(d=>{r(f=>{const p=f+d;return p.length>xc?p.slice(p.length-xc):p})},[]);return P.useEffect(()=>{u.current=!0;const f=`${window.location.protocol==="https:"?"wss:":"ws:"}//${window.location.host}/ws/aqueducts/${encodeURIComponent(e)}/peek`,p=new WebSocket(f);return a.current=p,p.onopen=()=>{u.current&&i(!0)},p.onmessage=x=>{u.current&&s(x.data)},p.onclose=()=>{u.current&&i(!1)},p.onerror=()=>{u.current&&i(!1)},()=>{u.current=!1,p.close(),a.current=null}},[e,s]),P.useEffect(()=>{o.current&&(o.current.scrollTop=o.current.scrollHeight)},[n]),y.jsxs("div",{className:"fixed inset-y-0 right-0 w-full md:w-[600px] bg-cistern-bg border-l border-cistern-border shadow-2xl z-50 flex flex-col",children:[y.jsxs("div",{className:"flex items-center justify-between px-4 py-3 border-b border-cistern-border",children:[y.jsxs("div",{className:"flex items-center gap-3",children:[y.jsx("h3",{className:"font-mono text-cistern-accent",children:e}),y.jsx("span",{className:"text-xs text-cistern-muted",children:"Peek"}),y.jsx("div",{className:`w-2 h-2 rounded-full ${l?"bg-cistern-green":"bg-cistern-red"}`})]}),y.jsx("button",{onClick:t,className:"text-cistern-muted hover:text-cistern-fg transition-colors text-lg leading-none",children:"×"})]}),y.jsx("pre",{ref:o,className:"flex-1 overflow-auto p-4 font-mono text-xs text-cistern-green bg-cistern-bg whitespace-pre-wrap break-all",children:n||"Connecting…"})]})}function qg(){const{data:e,error:t}=ep(),[n,r]=P.useState(null),l=P.useMemo(()=>{const o={};if(e)for(const a of e.flow_activities)o[a.droplet_id]=a;return o},[e]);if(t&&!e)return y.jsx("div",{className:"flex items-center justify-center h-full",children:y.jsxs("div",{className:"text-center",children:[y.jsx("div",{className:"text-cistern-red text-lg font-mono mb-2",children:"Connection Error"}),y.jsx("div",{className:"text-cistern-muted text-sm",children:t.message})]})});if(!e)return y.jsx("div",{className:"flex items-center justify-center h-full",children:y.jsx("div",{className:"text-cistern-muted font-mono",children:"Loading dashboard…"})});const i=new Set(e.cataractae.filter(o=>o.droplet_id).map(o=>o.droplet_id));return y.jsxs("div",{className:"flex-1 overflow-y-auto p-4 md:p-6 space-y-6",children:[y.jsx(bg,{cataractae:e.cataractae,flowingIds:i,activityMap:l,onPeek:r}),y.jsx(ey,{data:e}),n&&y.jsx(Zg,{aqueductName:n,onClose:()=>r(null)})]})}function bg({cataractae:e,flowingIds:t,activityMap:n,onPeek:r}){const l=e.filter(o=>o.droplet_id&&t.has(o.droplet_id)),i=e.filter(o=>!o.droplet_id);return y.jsxs("section",{children:[y.jsx("h2",{className:"text-sm font-mono text-cistern-muted uppercase tracking-wider mb-3",children:"Aqueducts"}),l.length>0&&y.jsx("div",{className:"space-y-3 mb-4",children:l.map(o=>y.jsx(wc,{cataractae:o,activity:n[o.droplet_id||""],isFlowing:!0,onPeek:r},o.name))}),i.length>0&&y.jsx("div",{className:"space-y-2",children:i.map(o=>y.jsx(wc,{cataractae:o,isFlowing:!1,onPeek:r},o.name))})]})}function ey({data:e}){return y.jsxs("div",{className:"grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4",children:[y.jsx(ty,{data:e}),y.jsx(ny,{items:e.cistern_items,blockedByMap:e.blocked_by_map}),y.jsx(ry,{items:e.pooled_items}),y.jsx(ly,{items:e.unassigned_items}),y.jsx(iy,{items:e.recent_items})]})}function ty({data:e}){return y.jsxs("div",{className:"bg-cistern-surface border border-cistern-border rounded-lg p-4",children:[y.jsx("h3",{className:"text-sm font-mono text-cistern-muted uppercase tracking-wider mb-3",children:"Cistern"}),y.jsxs("div",{className:"space-y-2",children:[y.jsxs("div",{className:"flex items-center justify-between",children:[y.jsx("span",{className:"text-sm text-cistern-fg",children:"Flowing"}),y.jsx("span",{className:"font-mono text-cistern-green",children:e.flowing_count})]}),y.jsxs("div",{className:"flex items-center justify-between",children:[y.jsx("span",{className:"text-sm text-cistern-fg",children:"Queued"}),y.jsx("span",{className:"font-mono text-cistern-yellow",children:e.queued_count})]}),y.jsxs("div",{className:"flex items-center justify-between",children:[y.jsx("span",{className:"text-sm text-cistern-fg",children:"Delivered"}),y.jsx("span",{className:"font-mono text-cistern-accent",children:e.done_count})]}),y.jsxs("div",{className:"flex items-center justify-between",children:[y.jsx("span",{className:"text-sm text-cistern-fg",children:"Pooled"}),y.jsx("span",{className:"font-mono text-cistern-red",children:e.pooled_items.length})]})]}),y.jsx("div",{className:"mt-3 pt-3 border-t border-cistern-border",children:y.jsxs("div",{className:"flex items-center gap-2",children:[y.jsx("div",{className:`w-2 h-2 rounded-full ${e.farm_running?"bg-cistern-green":"bg-cistern-muted"}`}),y.jsxs("span",{className:"text-xs text-cistern-muted",children:["Castellarius ",e.farm_running?"Running":"Stopped"]})]})})]})}function ny({items:e,blockedByMap:t}){return e.length===0?null:y.jsxs("div",{className:"bg-cistern-surface border border-cistern-border rounded-lg p-4",children:[y.jsxs("h3",{className:"text-sm font-mono text-cistern-muted uppercase tracking-wider mb-3",children:["Queue (",e.length,")"]}),y.jsx("div",{className:"space-y-1 max-h-64 overflow-y-auto",children:e.map(n=>y.jsx(Gi,{droplet:n,blockedBy:t[n.id]},n.id))})]})}function ry({items:e}){return e.length===0?null:y.jsxs("div",{className:"bg-cistern-surface border border-cistern-border rounded-lg p-4",children:[y.jsxs("h3",{className:"text-sm font-mono text-cistern-red uppercase tracking-wider mb-3",children:["Pooled (",e.length,")"]}),y.jsx("div",{className:"space-y-1 max-h-64 overflow-y-auto",children:e.map(t=>y.jsx(Gi,{droplet:t},t.id))})]})}function ly({items:e}){return e.length===0?null:y.jsxs("div",{className:"bg-cistern-surface border border-cistern-yellow/30 rounded-lg p-4",children:[y.jsxs("h3",{className:"text-sm font-mono text-cistern-yellow uppercase tracking-wider mb-3 flex items-center gap-2",children:[y.jsx("span",{children:"⚠"})," Orphaned (",e.length,")"]}),y.jsx("div",{className:"space-y-1 max-h-64 overflow-y-auto",children:e.map(t=>y.jsx(Gi,{droplet:t},t.id))})]})}function iy({items:e}){return e.length===0?null:y.jsxs("div",{className:"bg-cistern-surface border border-cistern-border rounded-lg p-4",children:[y.jsx("h3",{className:"text-sm font-mono text-cistern-muted uppercase tracking-wider mb-3",children:"Recent Flow"}),y.jsx("div",{className:"space-y-1 max-h-64 overflow-y-auto",children:e.slice(0,5).map(t=>y.jsx(Gi,{droplet:t},t.id))})]})}function Mr({title:e}){return y.jsx("div",{className:"flex items-center justify-center h-full",children:y.jsxs("div",{className:"text-center",children:[y.jsx("h2",{className:"text-lg font-mono text-cistern-fg mb-2",children:e}),y.jsx("p",{className:"text-sm text-cistern-muted",children:"Coming soon"})]})})}const oy=mg([{path:"/app",element:y.jsx(Wg,{}),children:[{index:!0,element:y.jsx(qg,{})},{path:"droplets",element:y.jsx(Mr,{title:"Droplets"})},{path:"castellarius",element:y.jsx(Mr,{title:"Castellarius"})},{path:"doctor",element:y.jsx(Mr,{title:"Doctor"})},{path:"logs",element:y.jsx(Mr,{title:"Logs"})},{path:"repos",element:y.jsx(Mr,{title:"Repos / Skills"})}]}]);Do.createRoot(document.getElementById("root")).render(y.jsx(Mc.StrictMode,{children:y.jsx(kg,{router:oy})}));
