+++ cmd/ct/assets/web/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Cistern</title>
-    <script type="module" crossorigin src="/app/assets/index-umFND8IE.js"></script>
+    <script type="module" crossorigin src="/app/assets/index-bUUc_DK2.js"></script>
     <link rel="stylesheet" crossorigin href="/app/assets/index-C_wjtcrj.css">
   </head>
   <body class="bg-cistern-bg text-cistern-fg">
