+++ web/vitest.config.ts
@@ -3,6 +3,9 @@ import react from '@vitejs/plugin-react';
 
 export default defineConfig({
   plugins: [react()],
+  define: {
+    'process.env.NODE_ENV': '"test"',
+  },
   test: {
     environment: 'jsdom',
     globals: true,
